+++ packages/terra-list/tests/jest/ListItem.test.jsx
@@ -1,40 +1,41 @@
 import React from 'react';
 import ThemeContextProvider from 'terra-theme-context/lib/ThemeContextProvider';
-
+/* eslint-disable-next-line import/no-extraneous-dependencies */
+import { shallowWithIntl, mountWithIntl } from 'terra-enzyme-intl';
 import ListItem from '../../src/ListItem';
 
 // Snapshot Tests
 it('should render with className', () => {
-  const shallowComponent = shallow(<ListItem className="textClass" />);
+  const shallowComponent = shallowWithIntl(<ListItem className="textClass" />).dive();
   expect(shallowComponent).toMatchSnapshot();
 });
 
 it('should render with content', () => {
-  const shallowComponent = shallow(<ListItem><p>text</p></ListItem>);
+  const shallowComponent = shallowWithIntl(<ListItem><p>text</p></ListItem>).dive();
   expect(shallowComponent).toMatchSnapshot();
 });
 
 it('should render with selected', () => {
-  const shallowComponent = shallow(<ListItem isSelectable isSelected />);
+  const shallowComponent = shallowWithIntl(<ListItem isSelectable isSelected />).dive();
   expect(shallowComponent).toMatchSnapshot();
 });
 
 it('should render with isSelectable', () => {
-  const shallowComponent = shallow(<ListItem isSelectable />);
+  const shallowComponent = shallowWithIntl(<ListItem isSelectable />).dive();
   expect(shallowComponent).toMatchSnapshot();
 });
 
 it('should render with hasChevron', () => {
-  const shallowComponent = shallow(<ListItem hasChevron />);
+  const shallowComponent = shallowWithIntl(<ListItem hasChevron />).dive();
   expect(shallowComponent).toMatchSnapshot();
 });
 
 it('should render with callback functions', () => {
   const mockCallBack = jest.fn();
 
-  const shallowComponent = shallow(
+  const shallowComponent = shallowWithIntl(
     <ListItem title="test" isSelectable onSelect={mockCallBack} refCallback={jest.fn()} />,
-  );
+  ).dive();
   expect(shallowComponent).toMatchSnapshot();
   shallowComponent.find('li').simulate('click');
   shallowComponent.find('li').simulate('keydown', { nativeEvent: { keyCode: 13 } });
@@ -43,7 +44,7 @@ it('should render with callback functions', () => {
 });
 
 it('correctly applies the theme context className', () => {
-  const wrapper = mount(
+  const wrapper = mountWithIntl(
     <ThemeContextProvider theme={{ className: 'orion-fusion-theme' }}>
       <ListItem />
     </ThemeContextProvider>,